task_01/requirements.txt
autopep8==1.6.0
dataclasses==0.8
decorator==5.1.1
mpmath==1.2.1
pkg-resources==0.0.0
pycodestyle==2.8.0
rel==0.4.7
six==1.16.0
sympy==1.9
toml==0.10.2
validators==0.18.2
websocket-client==1.3.1
